feat: added support for bare repositories (fixes #83)
A new option has been added to config: bare_repos. It is a list of pairs { git_dir, git_work_tree }.
If the plugin is used in a buffer which belongs to a bare repository, after the first call to git (`git rev-parse`)
which returns error since buffer is not in a git repo, subsequent similar calls will be made with the `--git-dir` and
`--work-tree` flags.
After the buffer has been found to be in one of the bare repositories the user added to the config the information is
cached for subsequent calls in the session